refactor: PathFinder domain을 interface으로 변경
- 외부 라이브러리에 의존하는 class인 경우 utils 패키지로 분리
